- core compatibility fixes
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@1742 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ windows/ListFilter.cpp
@@ -272,7 +272,7 @@ void ListFilter::columnChanged(bool doFilter) {
 		auto n = method.GetCount();
 		size_t col = getColumn();
 
-		for (size_t i = StringMatch::METHOD_LAST; i < n; ++i) {
+		for (int i = StringMatch::METHOD_LAST; i < n; ++i) {
 			method.DeleteString(StringMatch::METHOD_LAST);
 		}
